+++ src/werkzeug/routing/map.py
@@ -336,7 +336,7 @@ class Map:
 
         scheme = env["wsgi.url_scheme"]
         upgrade = any(
-            v.strip() == "upgrade"
+            v.strip(" \t") == "upgrade"
             for v in env.get("HTTP_CONNECTION", "").lower().split(",")
         )