+++ src/werkzeug/routing/matcher.py
@@ -143,13 +143,14 @@ class StateMachineMatcher:
                             remaining = [""]
 
                     converter_groups = sorted(
-                        match.groupdict().items(), key=lambda entry: entry[0]
+                        (
+                            item
+                            for item in match.groupdict().items()
+                            if item[0].startswith("__werkzeug_")
+                        ),
+                        key=lambda item: int(item[0][11:]),
                     )
-                    groups = [
-                        value
-                        for key, value in converter_groups
-                        if key[:11] == "__werkzeug_"
-                    ]
+                    groups = [item[1] for item in converter_groups]
                     rv = _match(new_state, remaining, values + groups)
                     if rv is not None:
                         return rv