+++ src/werkzeug/sansio/response.py
@@ -296,7 +296,7 @@ class Response:
         ct = self.headers.get("Content-Type")
 
         if ct:
-            return ct.split(";")[0].strip()
+            return ct.partition(";")[0].strip()
         else:
             return None