+++ src/werkzeug/serving.py
@@ -96,7 +96,10 @@ if t.TYPE_CHECKING:
 
 
 class DechunkedInput(io.RawIOBase):
-    """An input stream that handles Transfer-Encoding 'chunked'"""
+    """An input stream that handles ``Transfer-Encoding: chunked``. Only
+    used by the dev server, which must not be used in production. A production
+    WSGI server will have its own robust, secure chunk handler.
+    """
 
     def __init__(self, rfile: t.IO[bytes]) -> None:
         self._rfile = rfile
@@ -108,7 +111,7 @@ class DechunkedInput(io.RawIOBase):
 
     def read_chunk_len(self) -> int:
         try:
-            line = self._rfile.readline().decode("latin1").removesuffix("\r\n")
+            line = self._rfile.readline(100).decode("latin1").removesuffix("\r\n")
             _len = _plain_int(line, 16)
         except ValueError as e:
             raise OSError("Invalid chunk header") from e
@@ -150,7 +153,7 @@ class DechunkedInput(io.RawIOBase):
             if self._len == 0:
                 # Skip the terminating newline of a chunk that has been fully
                 # consumed. This also applies to the 0-sized final chunk
-                terminator = self._rfile.readline()
+                terminator = self._rfile.readline(2)
                 if terminator not in (b"\n", b"\r\n", b"\r"):
                     raise OSError("Missing chunk terminating newline")
 
@@ -448,17 +451,17 @@ class WSGIRequestHandler(BaseHTTPRequestHandler):
         msg = msg.translate(self._control_char_table)
         code = str(code)
 
-        if code[0] == "1":  # 1xx - Informational
+        if code.startswith("1"):  # 1xx - Informational
             msg = _ansi_style(msg, "bold")
         elif code == "200":  # 2xx - Success
             pass
         elif code == "304":  # 304 - Resource Not Modified
             msg = _ansi_style(msg, "cyan")
-        elif code[0] == "3":  # 3xx - Redirection
+        elif code.startswith("3"):  # 3xx - Redirection
             msg = _ansi_style(msg, "green")
         elif code == "404":  # 404 - Resource Not Found
             msg = _ansi_style(msg, "yellow")
-        elif code[0] == "4":  # 4xx - Client Error
+        elif code.startswith("4"):  # 4xx - Client Error
             msg = _ansi_style(msg, "bold", "red")
         else:  # 5xx, or any other response
             msg = _ansi_style(msg, "bold", "magenta")