+++ src/werkzeug/wsgi.py
@@ -145,8 +145,10 @@ def get_input_stream(
     safe_fallback: bool = True,
     max_content_length: int | None = None,
 ) -> t.IO[bytes]:
-    """Return the WSGI input stream, wrapped so that it may be read safely without going
-    past the ``Content-Length`` header value or ``max_content_length``.
+    """Return the WSGI input stream, wrapped so that it may be read safely
+    without going past the ``Content-Length`` header value or
+    ``max_content_length``. This must be used to pass a safe stream to other
+    functions, which are not responsible for enforcing this limit.
 
     If ``Content-Length`` exceeds ``max_content_length``, a
     :exc:`RequestEntityTooLarge`` ``413 Content Too Large`` error is raised.
@@ -559,7 +561,11 @@ class LimitedStream(io.RawIOBase):
                     return 0
 
                 if out_size:
-                    b[:out_size] = temp_b
+                    # Need to slice both sides. Math here is complicated.
+                    # size=10, remaining=5, out_size=3, remaining > out_size
+                    # Without slicing temp_b (b[:s] = tb), would try to
+                    # put 5 bytes into 3 bytes, causing b to resize to 12.
+                    b[:out_size] = temp_b[:out_size]
         else:
             # WSGI requires that stream.read is available.
             try: